Thanks for open sourcing this package. I made some performance improvements in case you're interested.

```
kevin@kevin-x220:~/code/go/src/github.com/kmanley/base36$ benchcmp old.txt new.txt 
benchmark             old ns/op     new ns/op     delta
BenchmarkEncode-4     2135          184           -91.38%
BenchmarkDecode-4     1543          928           -39.86%

benchmark             old allocs     new allocs     delta
BenchmarkEncode-4     13             1              -92.31%
BenchmarkDecode-4     1              0              -100.00%

benchmark             old bytes     new bytes     delta
BenchmarkEncode-4     112           16            -85.71%
BenchmarkDecode-4     16            0             -100.00%
```