@martinlindhe 2016-06-07
Very cool, thanks for this!